refactor(profile)!: restructure components & utilities
- Move ProfileCard from profile/page.tsx to components/profile/ProfileCard.tsx
- Extract phone formatting logic to lib/phone.ts
- Update imports and component organization
- Fix card height to prevent overflow issues

BREAKING CHANGE: ProfileCard component import path has changed
